move runners & reporters, extra common reporter logic
